COM-3408 - fix reviews
