+++ src/core/components/table/AttendanceTable/Composables/Attendances.js
@@ -1,33 +1,34 @@
 import { ref, computed } from 'vue';
+import { useStore } from 'vuex';
 import { useQuasar } from 'quasar';
 import { useRouter } from 'vue-router';
-import get from 'lodash/get';
-import isEmpty from 'lodash/isEmpty';
 import useVuelidate from '@vuelidate/core';
+import { required } from '@vuelidate/validators';
 import Attendances from '@api/Attendances';
-import Users from '@api/Users';
 import {
   HH_MM,
   DAY_OF_WEEK_SHORT,
   DAY_OF_MONTH,
   MONTH_SHORT,
+  COACH_ROLES,
+  VENDOR_ADMIN,
+  TRAINING_ORGANISATION_MANAGER,
 } from '@data/constants';
 import { upperCaseFirstLetter, formatIdentity, sortStrings, formatAndSortIdentityOptions } from '@helpers/utils';
 import { minArrayLength } from '@helpers/vuelidateCustomVal';
 import CompaniDate from '@helpers/dates/companiDates';
-import { required } from '@vuelidate/validators';
 import { NotifyPositive, NotifyNegative, NotifyWarning } from '@components/popup/notify';
 
-export const useAttendances = (course, isClientInterface, canUpdate, loggedUser) => {
+export const useAttendances = (course, isClientInterface, canUpdate, loggedUser, modalLoading) => {
   const $q = useQuasar();
   const $router = useRouter();
+  const $store = useStore();
 
   const attendances = ref([]);
   const traineeAdditionModal = ref(false);
   const newTraineeAttendance = ref({ trainee: '', attendances: [] });
   const potentialTrainees = ref([]);
   const loading = ref(false);
-  const modalLoading = ref(false);
 
   const attendanceRules = computed(() => ({
     newTraineeAttendance: { trainee: { required }, attendances: { minArrayLength: minArrayLength(1) } },
@@ -94,6 +95,13 @@ export const useAttendances = (course, isClientInterface, canUpdate, loggedUser)
 
   const disableCheckbox = computed(() => loading.value || !canUpdate.value || !!course.value.archivedAt);
 
+  const clientRole = computed(() => $store.getters['main/getClientRole']);
+
+  const vendorRole = computed(() => $store.getters['main/getVendorRole']);
+
+  const canAccessLearnerProfile = computed(() => COACH_ROLES.includes(clientRole.value) ||
+    [VENDOR_ADMIN, TRAINING_ORGANISATION_MANAGER].includes(vendorRole.value));
+
   const traineesCount = slotId => attendances.value.filter(a => a.courseSlot === slotId).length;
 
   const attendanceCheckboxValue = (traineeId, slotId) => {
@@ -101,20 +109,6 @@ export const useAttendances = (course, isClientInterface, canUpdate, loggedUser)
       return !!attendances.value.find(a => a.trainee === traineeId && a.courseSlot === slotId);
     }
     return false;
-  };
-
-  const getPotentialTrainees = async () => {
-    try {
-      let query;
-
-      if (isClientInterface) query = { companies: get(loggedUser.value, 'company._id') };
-      else query = { companies: course.value.companies.map(c => c._id) };
-
-      potentialTrainees.value = !isEmpty(query.companies) ? Object.freeze(await Users.learnerList(query)) : [];
-    } catch (error) {
-      potentialTrainees.value = [];
-      console.error(error);
-    }
   };
 
   const refreshAttendances = async (query) => {
@@ -230,7 +224,6 @@ export const useAttendances = (course, isClientInterface, canUpdate, loggedUser)
     newTraineeAttendance,
     potentialTrainees,
     loading,
-    modalLoading,
     // Computed
     attendanceColumns,
     traineesWithAttendance,
@@ -238,10 +231,10 @@ export const useAttendances = (course, isClientInterface, canUpdate, loggedUser)
     courseHasSlot,
     traineeFilterOptions,
     disableCheckbox,
+    canAccessLearnerProfile,
     // Methods
     traineesCount,
     attendanceCheckboxValue,
-    getPotentialTrainees,
     refreshAttendances,
     updateAttendanceCheckbox,
     slotCheckboxValue,
